pyproject.toml
[tool.poetry]
name = "noktech-deploy"
version = "0.1.2"
description = "Advanced deployment client with multiple protocols support"
authors = ["Brendown Ferreira"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
inquirer = "^3.1.3"
paramiko = "^3.4.0"
watchdog = "^6.0.0"
aiofiles = "^24.1.0"
aioftp = "^0.23.1"
pyyaml = "^6.0.1"
colorama = "^0.4.6"
jinja2 = "^3.1.2"
asyncssh = "^0.23.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
black = "^24.10.0"
ruff = "^0.8.2"
types-requests = "^2.32.0.20241016"
types-pyinstaller = "^6.11.0.20241028"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py310"
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "B", "S"]
ignore = ["S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.isort]
combine-as-imports = true
force-single-line = false
known-first-party = ["src"]

[tool.pytest]
asyncio_mode = "strict"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src --cov-report=html"
markers = [
    "asyncio: mark test as async",
] 

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
disallow_any_generics = false
disallow_incomplete_defs = false
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true
 